Some cleanup and moved Signer tests into their own file.
