+++ src/Signer/TimestampSigner.php
@@ -45,20 +45,13 @@ class TimestampSigner extends Signer {
         $timestamp = array_pop($exploded);
         $value = implode($this->sep, $exploded);
 
-        try {
-            $timestamp = bytes_to_int(base64_decode_($timestamp));
-        // @codeCoverageIgnoreStart
-        // not sure if this try catch is even needed - the builtins
-        // involved don't throw like their counter parts in python
-        } catch (Exception $ex) {
-            $timestamp = null;
-        }
-        // @codeCoverageIgnoreEnd
+        $timestamp = bytes_to_int(base64_decode_($timestamp));
 
         # Signature is *not* okay.  Raise a proper error now that we have
         # split the value and the timestamp.
-        if (!is_null($sig_err))
+        if (!is_null($sig_err)) {
             throw new BadTimeSignature((string) $sig_err, $value, $timestamp);
+        }
 
         # Signature was okay but the timestamp is actually not there or
         # malformed.  Should not happen, but well.  We handle it nonetheless
@@ -67,15 +60,17 @@ class TimestampSigner extends Signer {
 
         if(!is_null($max_age)) {
             $age = $this->get_timestamp() - $timestamp;
-            if($age > $max_age)
+            if($age > $max_age) {
                 throw new SignatureExpired(
                     "Signature age $age > $max_age seconds",
                     $value,
                     $this->timestamp_to_datetime($timestamp));
+            }
         }
 
-        if($return_timestamp)
+        if($return_timestamp) {
             return array($value, $this->timestamp_to_datetime($timestamp));
+        }
         return $value;
     }
